New object storage API (#2)
The included design document has the details, but this replaces the old Kate object storage API with a brand new one for the following reasons:

  - The need to version storage: when a player upgrades from one cartridge to another, this should be done without corrupting the data. Because Kate has no idea what's in the storage, it's schemaless, it's only the cartridge itself that can decide how migrations are to be performed.
  - The need for a more grounded model: this implementation comes with a small formal model and semantics for the object storage. The security properties are derived from that.

Some much needed refactoring of the IPC module also included. All untusted messages are now parsed at sources.

The migration protocol described in the document is not implemented by this patch, but will rather be implemented in a future version.