Support generating pre-built binaries and add more extensive installation instructions. (#3)
This includes the trade-offs in practicality and security that each
alternative installation mode offers. In the future Kate will (as much
as possible) be released as a fully sandbox application, but the
current pre-built binaries run unsandboxed.

It's likely that future versions will not be based on Electron in order
to achieve this, but that still leaves some problems with packaging
for Linux, as sandboxing there doesn't mash well with Chromium's own
sandboxing requirements.