New LJT VM for cartridge binary serialisation (#4)
The scary numbers in this PR are mostly the new generated schemas.

With this PR I'm moving LJT from its old AOT codegen approach, where the compiler would generate binary encoders and decoders for each language target, to a new VM-based approach. The compiler will now only generate a schema program for this VM (encoded in JSON format), as far as encoding/decoding goes. For untyped languages the generated schema is all you need to decode things. Encoding is a bit more awkward as LJT requires metadata to be included in the values that will be encoded as well.

The compiler is maintained in the Crochet repository so no changes here. But there are two new backends: json generates the AST for the codec VM, and ts-types generates type definitions and constructors for TypeScript. Kate uses both, and they're fairly sizable files, which is why there are so many line additions in this patch.

The VM is not optimised and there are no plans to optimise it soon, but there are plans to improve the error tracking and reporting. There's no noticeable performance regressions when encoding/decoding cartridges compared to the AOT codegen schemas.

There are two important additions to the schemas in LJT in this patch:

proper schema headers: a magic constant that's used by the parser to decide which parser version to pick, and a version number used by the parser to decide if the file can be parsed at all; LJT schemas are always backwards compatible, but not forwards compatible since they don't encode enough information in the serialisation;
versioning: although it's not possible to patch types in implementation (yet), the version is encoded in the serialisation. This will make it possible to add patching in the next release without breaking past cartridges, so this should be the last cartridge parsing fork in the code.
Changes are backwards compatible, however legacy cartridge formats (v2) are deprecated with this patch and will be removed on the first Kate stable release.