IndexedDB bridge (#7)
This patch implements a very small subset of the IndexedDB API, but does not follow the same specified semantics. In its current incantation it's enough to support Emscripten's IDBFS (which Ren'Py uses), but more work might be necessary if other engines use IndexedDB in different ways. Hopefully a complete replication of the IndexedDB API shan't be necessary :'>