Support more case modes (#9)
This patch adds a TV mode and a Full-screen mode, where the emulator does not provide virtual buttons and supports a higher screen resolution. These are meant for cases where the Kate emulator is run on a desktop/laptop computer or when attached to an external display/TV, and require a keyboard or gamepad.

For the Electron-based native version there's a `--fullscreen` flag that can be provided when running the application to run it in fullscreen. Otherwise the screen will just be maximised/unmaximised. Unfortunately the transparent frames have a handful of limitations that make this dynamic transition between fullscreen/not fullscreen modes a bit trickier.

For the single-cartridge version, the `kate-dist` tool now takes additional options to choose which case mode the emulator should be launched with.