Storage visualisation and management (#11)
This patch includes a new settings screen that allows users to manage cartridge and its related data, as well as giving a summary of how Kate is using the device's storage. There's a few new checks for storage health and some padding in the browser reported quota values, as at least Chromium will still allow the transaction to be committed even when it exceeds the quota, but subsequent write transactions will be aborted—this would e.g.: make the "Archive Cartridge" feature unusable, because even if the files deleted bring the storage under the quota, the new write at the end of the transaction will cause Chromium to roll it back.

There's also a new "archived cartridge" mode. This allows cartridge metadata to stick around in the console while removing all of its files to save space, and in turn simplifies a lot of the relations between tables (e.g.: we always expect a cartridge to exist if save data exists for that id, but we don't need the cartridge files for this).

This patch also makes Kate a bit more proactive in communicating to users when an installation or store operation would cause the underlying storage to run out of free space. This is done both before commiting large transactions and more actively in the background in case things change in the storage outside of Kate's control. The new "low storage" resource icon will be there to tell users they should be managing their storage more actively.

There's a few storage-unrelated bug fixes, the addition of deleting deprecated object stores, and some much needed reorganisation some api-related modules.