Move documentation to Sphinx (#12)
In preparing for a stable release, this patch moves all previous markdown documentation to a more properly organised collection of books using Sphinx. In the process some of the previously documented content was expanded upon, but the documentation is not final and still needs a lot of work. I haven't decided what I want to do with the glossary yet.