Refresh worker cache (#13)
This patch adds a new button in the Diagnostics & Recovery screen to refresh the worker cache, since by default it has no updating policy. We also refresh the cache automatically during a version upgrade or downgrade now, as those happen manually in Kate.

The missing piece is keeping the app code, the worker, and the CSS/etc resources pointing to the correct version, since Kate currently only versions the Kernel/OS JavaScript code. This will be done before the first stable release, since a reasonable maintenance policy needs to be defined first.