Breaking change: new cartridge format (#14)
This finally implements the cartridge changes that had been on the backlog for a while. It also removes code for parsing earlier versions of the cartridge format. The older formats had no real usage and thus don't need to be supported. The format introduced here must be supported for the whole life of Kate.

The patch is not complete—there are a few pending changes required for other in-progress patches (new capabilities and the store). But the format will not really change significantly. The limit for cartridge sizes was bumped to 1.4GB, but that still needs to be fully loaded in memory with the current implementation. Support for streaming will be added directly in LJT.

Note that there's no intended upgrade path from v0.26 to v0.28. Kate does not currently have any user base, so this is fine, but it means updating to 0.28 requires first deleting the entire Kate database and starting fresh.