Support requests to navigating to external URLs (#15)
This is a very initial patch to allow cartridges to open external URLs (in a similar fashion to `shell.open(url)` APIs that other platforms offer). Here we take some care to inform the user of the security and privacy implications of that action, and give them some information about the destination so they can make some risk assessment. However, due to the nature of Kate's cartridges, web browsers, and the internet, actual informed consent is not possible, and the threat of data exfiltration cannot be fully mitigated.

This patch also adds initial support for contextual capabilities. There will be separate patches to allow viewing and managing these capabilities; currently they're just auto-granted when the cartridge is installed.

There will also be future patches to implement proper auditing and quota supervising of the feature, that are specified here but not implemented yet.