Support managing capabilities (#16)
This patch includes a design document describing the general approach taken in Kate's capability security implementation. There's a lot of work that needs to happen in this area still, so treat this as a big initial draft. 'Sides the design documents, this patch also includes a minor settings screen for managing capabilities granted to a cartridge.