Add content rating and release type banners to cartridges in cases where more attention is needed (#17)
This adds some indicators about what to expect when playing a cartridge when they might require some attention. This includes a top-banner when the cartridge is not a regular release (e.g.: early access or demos would get a banner indicating this), and an age indication rating in the bottom.