Introduce a fairness supervisor (#20)
This patch introduces a "fairness supervisor". Right now this is pretty straight-forward, and just checks that a cartridge is not abusing modal dialogs to prevent users from using the device (since modal dialogs always stack on top of the current foreground screen). This means that from now on cartridges will only be able to show a modal dialog if they are the screen on foreground and if they are not showing another modal dialog (it shouldn't be possible with the first restriction in any case).

There may be more fairness quotas applied later, particularly once Kate supports multiple processes, as there we'd like to ensure that no cartridge hogs the entire device to itself.