Change licences from MIT to MPL v2 (#22)
This changes the existing source code licence from MIT to MPL v2. The primary reason for it is to improve the guarantees that end-users will be able to audit the software they're given even if they do not download it from a the official github repository. Since MIT does not require source code to be divulged, or even that modifications to source code be stated, it does not fulfil that purpose.

Transparency is a core tenet of Kate's security (it is one of the reasons why all features must have proper formal semantics and a public threat model). While we cannot ensure that malicious actors will comply with the licence, we can at least mitigate some of the damage from changes that, while not malicious, still change the threat models that we communicate to end users and that they expect from Kate. At the very least, the MPL encourages people to be open about their changes, and Kate's informational text surrounding the licences now strongly encourages this as well.

Relicencing it under the GPL v3 was considered. However, since the kate-dist tool needs to embed the whole OS with the cartridge for standalone distribution purposes, that complicated matters a bit. Imposing Kate's licence and guarantees on external releases is a non-goal, and we already state that no security guarantees can be provided by us in those cases. Going with the MPL should break no existing uses that were served by the MIT licence in practice, and supported OS modifications are still a work in progress in any case.

The new licence applies from the next Kate release (so v0.23.9+).