New Kate controls (#23)
As part of moving from the current prototype state into a more stable release, this patch pins down the controls for Kate, which are now all digital. A new Berry and Sparkle buttons have been added, as many games and tools needed a little bit of extra input---as a result the Menu button is no longer dual-use, but rather remains at disposal of the cartridge. The new Berry button assumes the role of system-specific functions.

The new buttons don't have a real button image yet, that'll be left for a future patch. There's also a new handheld case in this patch that works better with mobile phones by being more economical on the horizontal space usage and letting the Kate screen fill more of the real screen. This helps make some VNs at least somewhat readable (though most VNs are designed for PC screens and have very tiny fonts anyway... a design choice unfortunately shared by many modern games in general).

The virtual console has been cleaned a little bit. Input-dealing routines are now moved into the new `input` module and are connected in less haphazardly ways. There are other things that need to be likewise given that treatment, but don't belong in this patch. By the release of 23.10 the kernel should be stable, but the OS portion will need quite a bit of love still.