Reserve space for virtual keyboards in text input.
Note that the virtual keyboard on steam deck is a bit special and not handled in the browser (which should also be the case for other regular-OS-level virtual keyboards), so in those cases the text input still has a higher chance of being covered by the virtual keyboard. At least on steam deck, the text input is luckily just slightly above the keyboard when Kate is placed in fullscreen mode.