Make keyboard input bridge more flexible (#33)
Different engines listen to keyboard events in different elements (e.g.: Bitsy uses document, Ren'py uses window, Godot uses the canvas it's rendering to), so the element we dispatch events to needs to be configurable.

This patch introduces this in a backwards compatible way. A new keyboard-input-proxy-v2 bridge works like input-proxy, but takes in an additional "selector" parameter, which can be "window", "document", or a CSS selector to find an element on the page (we poll for up to 60 seconds). Uses of the legacy input-proxy bridge are converted to the new one at runtime, but parsing support for the older value will always be kept.