Make process handling less ad-hoc (#34)
Continuing with the (long) move away from the early prototype code into properly designed code, this makes process handling in the Kate kernel more principled. The added design document goes into details, but the TL;DR is that this introduces a new "Process" and "ProcessManager" object in the kernel layer. These objects are responsible for managing all process communication with the underlying sandboxed iframe running the actual cartridge code.

Notably, this patch adds support for the possibility of running multiple processes in Kate, however the OS front-end does not yet make use of that capability. This will be used in the future to run services, drivers, and other background processes, but how to best do that still needs some discovery, since background processes bring with them a whole host of new security and privacy threats.

This patch also improves the process communication channel a little bit; moving away from the globally shared window channel to a per-process MessageChannel (which is negotiated initially on the shared window channel). This simplifies routing of those messages considerably.