Move media storage to new partitioned file storage (#36)
Continuing the move from giant blobs in IndexedDB to OPFS-backed
partitioned file storage, this moves media files to the new partitioned
file storage. Note that we cannot yet relax the restrictions on video
recording sizes because we require all chunks to fit in memory. We may
relax things by writing chunks to a temporary bucket instead of holding
them in memory --- this would also open up the possibility of doing
higher quality recording.