Temporary buckets (#39)
This patch allows processes to store temporary files in buckets up to a
configurable size. The new `store-temporary-files` passive capability
decides the "maximum amount of storage" configuration for each process,
which is otherwise capped at 60% of the available storage.

Importer is now written using the temporary files, which improves its
memory usage a little bit.