Allow "synchronous" loading of cartridge files (#41)
In general, cartridge files live only in the kernel side, and cartridge
processes must use an asynchronous RPC call to read them. This generally
works for games made exclusively for Kate, however existing engines have
varied needs of reading data "synchronously" (at least from the
perspective of the game's code), so we need to support that as well.

This patch adds a way of pre-loading certain assets at the time the
cartridge sandbox is being built, such that all reads of these assets
from the cartridge's process appear to be synchronous (since they're
moved to the cartridge's process at the time the sandbox is built). This
in turn lets us support games written in RPG Maker MV and Godot 3 and 4
with no changes required on the game's code. The environment just
fulfils all of the game's code expectations instead.

Note that Godot 4 games are not yet *enabled* in Kate due to the
security vulnerability described in
https://github.com/orgs/qteatime/discussions/1 --- cartridge processes
need to be fully isolated before Godot 4 games can be enabled, as their
processing requirements make it possible for cartridges to exploit
Spectre vulnerabilities to read privileged information from the Kate
kernel process, at least in theory.