Release previous cartridge bucket on version overwrite/update.
This mostly accounts for the case where developer version overwrites (installing a new cartridge that's the same version as an existing one) lead to unbounded storage growth because bucket keys are bound to a cartridge/version, so the new cartridge holds a strong reference to all previous cartridge buckets it replaced as well and those will only be collected when the cartridge is upgraded or archived.

This also improves storage usage for regular upgrades a little bit by making sure the garbage collection of the previous bucket kicks in faster.