Add new Publisher application (#42)
The Publisher is a new Kate-cartridge-based application for creating
Kate cartridges. The main advantages over the command-line-based `kart`
application right now are that the Publisher is fully sandboxed (using
the same sandboxing as other Kate cartridges), and it's hooked to the
Kate key store to automatically sign cartridges with the appropriate
ECDSA key associated with the user's domain.

This is a pretty bare-bones implementation of the Publisher, though. The
configuration is still provided as a JSON file following the `kart`
configuration. There's a rich GUI for describing the cartridge's
configuration coming in the next release, but it requires a fair amount
of effort so I'm skipping it for this one.

The Publisher is the recommended way of creating cartridges for Kate,
but the command line applications will be kept *for strict automation
purposes* until the Headless Kate feature is implemented (which will
allow interacting with cartridges programmatically through typed APIs
while being subject to the same sandboxing and capability security
expectations as regular GUI cartridges).

Note that there are some kernel changes rolled up in this one that were
required to expose some new features required by the Publisher.