Add install prompts before cartridge installation (#45)
This patch adds a small install prompt before the actual cartridge
installation. The prompt provides information about what cartridge is
being installed, whether the cartridge signature has been verified, and
any higher-risk capabilities that the cartridge requires to work.

There's still some work that needs to happen on both the
capability-granting side and on showing what has been verified about the
cartridge. Those will have to be done in a separate patch as it requires
a bit more of deliberation; the design in this case is straightforward
enough to not have an accompanying design document, but signature
verification and communication will need one eventually.