Change licence of core components to GPLv2 (or later) (#46)
This was considered already in a previous pull request (see #22), but OS
extension cartridges weren't properly considered back then since their
specific design was not very well formed, and the MPL does not provide
the necessary protections for them. This is a problem for some of the
items in the roadmap (e.g.: cartridges that provide OS extensions).

For example, suppose someone releases a cartridge that provides a new OS
theme to make everything look like a pastel-coloured pixel art game of
certain franchise. Such cartridge could change the visual behaviour of
all of the Kate operating system, but wouldn't require any of the
commonly powerful capabilities (e.g.: "passive network access"), and
still it could cause a lot of damage to Kate's security guarantees.

Much of the security guarantees we provide for third-party cartridges
ride on the [trust and resource indicator
bar](https://docs.kate.qteati.me/user/manual/security/trust.html). In
Kate code there's special code to make sure this bar communicates enough
information on whether the process is trusted (part of the OS) or not. A
theme has no obligation of maintaining the same, and there's no form of
static analysis we can employ to check either way. Whether it would be
an honest mistake (additional effort is required to make the visual
trusted/untrusted distinction) or a deliberate omission, the end result
is the same: players would not be able to distinguish which process
screens they can trust to carry the security and privacy guarantees Kate
communicates. As a result, users could end up making dangerous decisions
while being misled by their OS.

This is a scenario we want to avoid at all costs, and so we require
changes that are made directly to the OS (and thus impact these security
and privacy guarantees) to also provide the source code and build
instructions such that users can, realistically, audit what they're
running if they feel inclined to. The MPL unfortunately doesn't provide
this guarantee---it only covers changes that are made directly to files
marked as MPL-covered, but cartridges will contain new files that are
linked dynamically through the upcoming file system layer and extension
feature.

The reason the previous PR avoided using GPL was, primarily, because
releasing things under GPL complicates things a lot for end users. Many
open source licences are incompatible with it, and even versions of the
GPL are incompatible with each other. Sadly this complication appears to
be unavoidable if we want direct OS modifications to be auditable, so
it's better to just go ahead and make the changes before the project is
publicly available :')

In the interest of making the GPL extend only to changes that would in
some way change the OS behaviour and feedback directly, the licencing
change is restricted in the following way:

- It only applies to the Kate OS itself (i.e.: `kate-core`, `kate-api`,
`kate-bridges`, and `kate-desktop`), the build system (`make.js`,
`support/*`), and things that are used to build cartridges
(`kate-tools`, `kate-importer`, and `kate-publisher`). Technically we
could avoid extending it to the cartridge-making tools, but it's good to
reassure developers exactly how their cartridge gets built and what gets
included in it, and there are no drawbacks to either developers or
end-users.

- There's a special exception with linking dynamically to the runtime
API (`kate-api` and `kate-bridges`) so the vast majority of cartridges
don't get accidentally covered by the GPL --- there's no need for them
to, we already treat all regular cartridges as untrusted so they can't
subvert the security/privacy expectations we set.